refactor: separated game logic and performance tests from the main file
