+++ types.h
@@ -1,0 +1,9 @@
+#pragma once
+
+typedef struct {
+    unsigned int width;
+    unsigned int height;
+
+    unsigned char *leftGrid;
+    unsigned char *rightGrid;
+} Universe;
